calorieapp/requirements.txt
openai>=1.0.0
flask>=2.3.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
flask-login>=0.6.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
requests>=2.31.0
scikit-learn>=1.3.0
chromadb>=0.4.0
Pillow>=10.0.0
werkzeug>=2.3.0
